HybridCloudApp/Code/Backend/MQTT_DB_Plugin_Service/requirements.txt
paho-mqtt
mysql-connector-python
PySocks